nitfix (Resolve nit observations) Resolve nit observations + walkthrough step 6 integration.
+ Remove unnecessary comments in pom.xml.

+ Changed _ prefix notation in the Comment.java constructor and fixed to a more idiomatic
way "this.attribute" to construct the object. Since attributes are never modified, the final
access modifier was added.

- NewCommentsServlet.java was refactored inside the CommentsServlet.java since both functionalities
were mergeable.

+ CommentsServlet.java now contains the NewCommentsServlet.java doPost method logic, including all the
hotfixes to the nit observations. The query now looks for a Query String limit parameter that can be
attached to limit the amount of comments to fetch from the datastore.

+ The DeleteCommentServlet.java offers the client a delete functionality to remove comments from the
App Engine Datastore. The servlet receives a comment id and tries to remove it.

+ Added required flag to textarea for the comment submission to enforce every comment to include a content.
Moved Network section to the rightmost column since it gives more balance to the page visual load. Hotfixed
function names convention from snake case to camel case.
Added a selection dropdown for the client to limit the amount of comments to load on the comments list.

+ Nitfix on queries/index.js from snake case to camel case convention.

+ Added deleteComment function in mutations/index.js to handle comment deletion triggering by the client.
The function requests the clicked comment to be deleted and removes its node from the list.

+ Added needed classes to style.css to style the new dropdown to limit amount of messages to be loaded.

*** No breaking changes ***