nitfix (Resolve) Resolve nitfixes + refactor portfolio order and view.
+ Resolve mentioned nits, and resolved issue in doPost POST body reading.
Since the body was not being extracted correctly, it was necessary to build
a parser function to extract the body of a request. Added as a return value
in the response of doPost, the new comment created.

+ Added delete functionality to webapp as per requirements. The DeleteCommentServlet.java
expects and id to search for the element to delete from the datastore.

+ Added /functions/index.js to store generic functions to construct HTML elements and
other functional snippets. createCommentElement function creates an element from an
incoming comment.

+ Refactored index.html structure to a more user friendly approach. Removed hobbies section
since it was decided to be removed and changed for a list of projects.

+ Added newCommentForm listener to handle a comment submission in requests/mutations/index.js
The form avoids restarting the webpage and handles the form submission by the client, and sends
it over to the server.

+ Limit variable nitfix + added attributes to getComments() function to add limit and pagination
to the new projects section. The getComments() function fetches a series of my Github repos.

-- Deleted hobbies.js and projects.js since functions were no longer being utilized in the
application. Removed references from /constants/index.js

+ Changed profile picture file.

+ Added some keyFrames and media queries to make the application more smooth and responsive.
Added and refactor servera CSS classes to better fit the new customization.

*** No breaking changes ***