nitfix (Resolve comments) Resolve nitfixes and code refactoring.
+ Add {emotion: String} attribute to Comment.java to store a descriptive
tag of the comment's sentiment.

+ Modified CommentsServlet.java to handle the new Comment.java emotion attribute
on querying and creating comments. The tag for new comments is defined in doPost.
Changed getSentiment() to analyzeSentiment() to avoid conflicting with getters
and setters naming convention.

+ Removed emotion tag creation in functions/index.js since tag is now set
on the backend service. This way the webapp remains agnostic to the business
logic.

+ Add new google-maps/index.js import in index.html.

+ Refactored the loadMap() function from queries/index.js to a new folder
google-maps/index.js to separate and group types of queries involving
separate API's.

*** No breaking changes ***